requirements.txt
requests
fastapi
uvicorn
bs4
jinja2
schedule
python-dotenv
Pillow
google-genai
python-multipart
